Create Palindrome
Palindrome Program in Java
Palindrome number in java: A palindrome number is a number that is same after reverse. For example 545, 151, 34543, 343, 171, 48984 are the palindrome numbers. It can also be a string like LOL, MADAM etc.
Let's see the palindrome program in java. In this java program, we will get a number variable and check whether number is palindrome or not.