Create Armstrong Number
Armstrong Number in Java
In this section, we will discuss what is Armstrong number and also create Java programs to check if the given number is an Armstrong number or not. The Armstrong number program frequently asked in Java coding interviews and academics.

Armstrong Number Java Program
The following Java program prints all the Armstrong numbers up to the specified limit.