Create Random Number Generation
Random Number Generation in Java 8
In Java 8, a new method ints() has been added to the Random class. We must import the java.util.Random before using the method.

ints():

The pseudorandom int values generated the same as calling the nextInt() method. It returns an unlimited stream of pseudorandom int values.

ints(long streamSize):

The method parses a parameter streamSize of type long. It specifies the number of values to be generated. The pseudorandom int values generated the same as calling the nextInt() method. It also returns a stream of randomly generated int values. It throws IllegalArgumentException if the stream size is less than zero.

ints(long streamSize, int randomNumberOrigin, int randomNumberBound):

Let's create a program that generates a stream of integers using the ints() method of the Random class.